+++ packages/core/src/components.d.ts
@@ -7,11 +7,13 @@
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { AccordionInternalState, AccordionSectionAnimationEndEvent, AccordionSectionToggleClickEvent, AccordionVariant } from "./components/accordion/accordion.interfaces";
 import { AccordionHeading, AccordionSectionState } from "./components/accordion/components/accordion-section.interfaces";
-import { AnnotationToggleEvent } from "./components/annotation-output/annotation-output.interfaces";
+import { AnnotationButtonClickEvent } from "./components/annotation-button/annotation-button";
+import { AnnotationOutputCloseEvent } from "./components/annotation-output/annotation-output";
 import { Suggestion } from "./components/autosuggest/autosuggest.interfaces";
 import { DsoCardClickedEvent, ImageShape } from "./components/card/card.interfaces";
 import { CardContainerMode } from "./components/card-container/card-container.interfaces";
 import { DsoDatePickerChangeEvent, DsoDatePickerDirection, DsoDatePickerFocusEvent, DsoDatePickerKeyboardEvent } from "./components/date-picker/date-picker.interfaces";
+import { DocumentComponentOpenToggleEvent, DocumentComponentToggleAnnotationEvent, DocumentComponentWijzigactie } from "./components/document-component/document-component";
 import { ExpandableAnimationEndEvent } from "./components/expandable/expandable";
 import { EditAction, ExpandableHeadingToggleEvent, HeadingTags } from "./components/expandable-heading/expandable-heading.interfaces";
 import { HeaderEvent, HeaderMenuItem } from "./components/header/header.interfaces";
@@ -32,11 +34,13 @@ import { TreeViewItem, TreeViewPointerEvent } from "./components/tree-view/tree-
 import { FilterpanelEvent, MainSize, ViewerGridChangeSizeEvent } from "./components/viewer-grid/viewer-grid.interfaces";
 export { AccordionInternalState, AccordionSectionAnimationEndEvent, AccordionSectionToggleClickEvent, AccordionVariant } from "./components/accordion/accordion.interfaces";
 export { AccordionHeading, AccordionSectionState } from "./components/accordion/components/accordion-section.interfaces";
-export { AnnotationToggleEvent } from "./components/annotation-output/annotation-output.interfaces";
+export { AnnotationButtonClickEvent } from "./components/annotation-button/annotation-button";
+export { AnnotationOutputCloseEvent } from "./components/annotation-output/annotation-output";
 export { Suggestion } from "./components/autosuggest/autosuggest.interfaces";
 export { DsoCardClickedEvent, ImageShape } from "./components/card/card.interfaces";
 export { CardContainerMode } from "./components/card-container/card-container.interfaces";
 export { DsoDatePickerChangeEvent, DsoDatePickerDirection, DsoDatePickerFocusEvent, DsoDatePickerKeyboardEvent } from "./components/date-picker/date-picker.interfaces";
+export { DocumentComponentOpenToggleEvent, DocumentComponentToggleAnnotationEvent, DocumentComponentWijzigactie } from "./components/document-component/document-component";
 export { ExpandableAnimationEndEvent } from "./components/expandable/expandable";
 export { EditAction, ExpandableHeadingToggleEvent, HeadingTags } from "./components/expandable-heading/expandable-heading.interfaces";
 export { HeaderEvent, HeaderMenuItem } from "./components/header/header.interfaces";
@@ -143,10 +147,13 @@ export namespace Components {
         /**
           * To link the Annotation Button with `aria-controls` to a different element, most likely an Annotation Output.
          */
-        "identifier": string;
+        "identifier": string | undefined;
+        /**
+          * Set to true when the annotation is open.
+         */
+        "open": boolean;
     }
     interface DsoAnnotationOutput {
-        "_toggleAnnotation": (e: MouseEvent | KeyboardEvent, identifier: string) => Promise<void>;
         /**
           * This text will be displayed above the annotation-output when opened
          */
@@ -155,6 +162,10 @@ export namespace Components {
           * The annotation-button that toggles this component should have the same identifier.
          */
         "identifier": string;
+        /**
+          * Set to `true` to show content.
+         */
+        "open": boolean;
     }
     interface DsoAttachmentsCounter {
         /**
@@ -285,6 +296,68 @@ export namespace Components {
           * Date value. Must be in Dutch date format: DD-MM-YYYY.
          */
         "value": string;
+    }
+    interface DsoDocumentComponent {
+        /**
+          * Enables annotations.
+         */
+        "annotated": boolean;
+        /**
+          * Marks as draft.
+         */
+        "draft": boolean;
+        /**
+          * Marks the Document Component as expired.
+         */
+        "expired": boolean;
+        /**
+          * Marks this Document Component as belonging to an active filter.
+         */
+        "filtered": boolean;
+        /**
+          * The heading element to use.
+         */
+        "heading": "h2" | "h3" | "h4" | "h5" | "h6";
+        /**
+          * The Inhoud XML.
+         */
+        "inhoud"?: string;
+        /**
+          * The Label XML.
+         */
+        "label"?: string;
+        /**
+          * When a child Document Component has a status "Draft".
+         */
+        "nestedDraft": boolean;
+        /**
+          * Marks this Document Component as not-applicable.
+         */
+        "notApplicable": boolean;
+        /**
+          * The Nummer XML.
+         */
+        "nummer"?: string;
+        /**
+          * This boolean attribute indicates whether the children are visible.
+         */
+        "open": boolean;
+        /**
+          * The Opschrift XML.
+         */
+        "opschrift"?: string;
+        /**
+          * Marks Document Component as reserved.
+         */
+        "reserved": boolean;
+        /**
+          * Type of Document Component.
+         */
+        "type"?: string;
+        /**
+          * The wijzigactie as in STOP.
+         */
+        "wijzigactie"?: DocumentComponentWijzigactie;
     }
     interface DsoDropdownMenu {
         /**
@@ -802,10 +875,16 @@ export namespace Components {
          */
         "overlayOpen": boolean;
     }
+    interface DsotDocumentComponentDemo {
+    }
 }
 export interface DsoAccordionSectionCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLDsoAccordionSectionElement;
+}
+export interface DsoAnnotationButtonCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLDsoAnnotationButtonElement;
 }
 export interface DsoAnnotationOutputCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -822,6 +901,10 @@ export interface DsoCardCustomEvent<T> extends CustomEvent<T> {
 export interface DsoDatePickerCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLDsoDatePickerElement;
+}
+export interface DsoDocumentComponentCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLDsoDocumentComponentElement;
 }
 export interface DsoExpandableCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -983,6 +1066,12 @@ declare global {
     var HTMLDsoDatePickerElement: {
         prototype: HTMLDsoDatePickerElement;
         new (): HTMLDsoDatePickerElement;
+    };
+    interface HTMLDsoDocumentComponentElement extends Components.DsoDocumentComponent, HTMLStencilElement {
+    }
+    var HTMLDsoDocumentComponentElement: {
+        prototype: HTMLDsoDocumentComponentElement;
+        new (): HTMLDsoDocumentComponentElement;
     };
     interface HTMLDsoDropdownMenuElement extends Components.DsoDropdownMenu, HTMLStencilElement {
     }
@@ -1158,6 +1247,12 @@ declare global {
         prototype: HTMLDsoViewerGridElement;
         new (): HTMLDsoViewerGridElement;
     };
+    interface HTMLDsotDocumentComponentDemoElement extends Components.DsotDocumentComponentDemo, HTMLStencilElement {
+    }
+    var HTMLDsotDocumentComponentDemoElement: {
+        prototype: HTMLDsotDocumentComponentDemoElement;
+        new (): HTMLDsotDocumentComponentDemoElement;
+    };
     interface HTMLElementTagNameMap {
         "dso-accordion": HTMLDsoAccordionElement;
         "dso-accordion-section": HTMLDsoAccordionSectionElement;
@@ -1173,6 +1268,7 @@ declare global {
         "dso-card": HTMLDsoCardElement;
         "dso-card-container": HTMLDsoCardContainerElement;
         "dso-date-picker": HTMLDsoDatePickerElement;
+        "dso-document-component": HTMLDsoDocumentComponentElement;
         "dso-dropdown-menu": HTMLDsoDropdownMenuElement;
         "dso-expandable": HTMLDsoExpandableElement;
         "dso-expandable-heading": HTMLDsoExpandableHeadingElement;
@@ -1202,6 +1298,7 @@ declare global {
         "dso-tooltip": HTMLDsoTooltipElement;
         "dso-tree-view": HTMLDsoTreeViewElement;
         "dso-viewer-grid": HTMLDsoViewerGridElement;
+        "dsot-document-component-demo": HTMLDsotDocumentComponentDemoElement;
     }
 }
 declare namespace LocalJSX {
@@ -1299,7 +1396,15 @@ declare namespace LocalJSX {
         /**
           * To link the Annotation Button with `aria-controls` to a different element, most likely an Annotation Output.
          */
-        "identifier": string;
+        "identifier": string | undefined;
+        /**
+          * Emitted when user activates the button.
+         */
+        "onDsoClick"?: (event: DsoAnnotationButtonCustomEvent<AnnotationButtonClickEvent>) => void;
+        /**
+          * Set to true when the annotation is open.
+         */
+        "open"?: boolean;
     }
     interface DsoAnnotationOutput {
         /**
@@ -1313,7 +1418,11 @@ declare namespace LocalJSX {
         /**
           * This event is emitted when the user activates the Annotation Button.
          */
-        "onDsoToggle"?: (event: DsoAnnotationOutputCustomEvent<AnnotationToggleEvent>) => void;
+        "onDsoClose"?: (event: DsoAnnotationOutputCustomEvent<AnnotationOutputCloseEvent>) => void;
+        /**
+          * Set to `true` to show content.
+         */
+        "open"?: boolean;
     }
     interface DsoAttachmentsCounter {
         /**
@@ -1468,6 +1577,76 @@ declare namespace LocalJSX {
           * Date value. Must be in Dutch date format: DD-MM-YYYY.
          */
         "value"?: string;
+    }
+    interface DsoDocumentComponent {
+        /**
+          * Enables annotations.
+         */
+        "annotated"?: boolean;
+        /**
+          * Marks as draft.
+         */
+        "draft"?: boolean;
+        /**
+          * Marks the Document Component as expired.
+         */
+        "expired"?: boolean;
+        /**
+          * Marks this Document Component as belonging to an active filter.
+         */
+        "filtered"?: boolean;
+        /**
+          * The heading element to use.
+         */
+        "heading"?: "h2" | "h3" | "h4" | "h5" | "h6";
+        /**
+          * The Inhoud XML.
+         */
+        "inhoud"?: string;
+        /**
+          * The Label XML.
+         */
+        "label"?: string;
+        /**
+          * When a child Document Component has a status "Draft".
+         */
+        "nestedDraft"?: boolean;
+        /**
+          * Marks this Document Component as not-applicable.
+         */
+        "notApplicable"?: boolean;
+        /**
+          * The Nummer XML.
+         */
+        "nummer"?: string;
+        /**
+          * Emitted when the user activates the annotation button.
+         */
+        "onDsoAnnotationToggle"?: (event: DsoDocumentComponentCustomEvent<DocumentComponentToggleAnnotationEvent>) => void;
+        /**
+          * Emitted when the user activates the toggle.
+         */
+        "onDsoOpenToggle"?: (event: DsoDocumentComponentCustomEvent<DocumentComponentOpenToggleEvent>) => void;
+        /**
+          * This boolean attribute indicates whether the children are visible.
+         */
+        "open"?: boolean;
+        /**
+          * The Opschrift XML.
+         */
+        "opschrift"?: string;
+        /**
+          * Marks Document Component as reserved.
+         */
+        "reserved"?: boolean;
+        /**
+          * Type of Document Component.
+         */
+        "type"?: string;
+        /**
+          * The wijzigactie as in STOP.
+         */
+        "wijzigactie"?: DocumentComponentWijzigactie;
     }
     interface DsoDropdownMenu {
         /**
@@ -2063,6 +2242,8 @@ declare namespace LocalJSX {
          */
         "overlayOpen"?: boolean;
     }
+    interface DsotDocumentComponentDemo {
+    }
     interface IntrinsicElements {
         "dso-accordion": DsoAccordion;
         "dso-accordion-section": DsoAccordionSection;
@@ -2078,6 +2259,7 @@ declare namespace LocalJSX {
         "dso-card": DsoCard;
         "dso-card-container": DsoCardContainer;
         "dso-date-picker": DsoDatePicker;
+        "dso-document-component": DsoDocumentComponent;
         "dso-dropdown-menu": DsoDropdownMenu;
         "dso-expandable": DsoExpandable;
         "dso-expandable-heading": DsoExpandableHeading;
@@ -2107,6 +2289,7 @@ declare namespace LocalJSX {
         "dso-tooltip": DsoTooltip;
         "dso-tree-view": DsoTreeView;
         "dso-viewer-grid": DsoViewerGrid;
+        "dsot-document-component-demo": DsotDocumentComponentDemo;
     }
 }
 export { LocalJSX as JSX };
@@ -2127,6 +2310,7 @@ declare module "@stencil/core" {
             "dso-card": LocalJSX.DsoCard & JSXBase.HTMLAttributes<HTMLDsoCardElement>;
             "dso-card-container": LocalJSX.DsoCardContainer & JSXBase.HTMLAttributes<HTMLDsoCardContainerElement>;
             "dso-date-picker": LocalJSX.DsoDatePicker & JSXBase.HTMLAttributes<HTMLDsoDatePickerElement>;
+            "dso-document-component": LocalJSX.DsoDocumentComponent & JSXBase.HTMLAttributes<HTMLDsoDocumentComponentElement>;
             "dso-dropdown-menu": LocalJSX.DsoDropdownMenu & JSXBase.HTMLAttributes<HTMLDsoDropdownMenuElement>;
             "dso-expandable": LocalJSX.DsoExpandable & JSXBase.HTMLAttributes<HTMLDsoExpandableElement>;
             "dso-expandable-heading": LocalJSX.DsoExpandableHeading & JSXBase.HTMLAttributes<HTMLDsoExpandableHeadingElement>;
@@ -2156,6 +2340,7 @@ declare module "@stencil/core" {
             "dso-tooltip": LocalJSX.DsoTooltip & JSXBase.HTMLAttributes<HTMLDsoTooltipElement>;
             "dso-tree-view": LocalJSX.DsoTreeView & JSXBase.HTMLAttributes<HTMLDsoTreeViewElement>;
             "dso-viewer-grid": LocalJSX.DsoViewerGrid & JSXBase.HTMLAttributes<HTMLDsoViewerGridElement>;
+            "dsot-document-component-demo": LocalJSX.DsotDocumentComponentDemo & JSXBase.HTMLAttributes<HTMLDsotDocumentComponentDemoElement>;
         }
     }
 }