+++ packages/core/src/components/annotation-output/annotation-output.interfaces.ts
@@ -1,4 +1,0 @@
-export interface AnnotationToggleEvent {
-  originalEvent?: MouseEvent | KeyboardEvent;
-  open: boolean;
-}