+++ packages/react/src/components/expandable/expandable.content.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 export const expandableContent: JSX.Element = (
-  <div slot="expandable-content" className="dso-rich-content">
+  <div className="dso-rich-content">
     <h3>Expandable</h3>
     <span>Dit is een expandable</span>
   </div>