+++ storybook/src/example-pages/snippets/document-component/document-component.ts
@@ -1,0 +1,8 @@
+import { html } from "lit-html";
+import { examplePageFactory } from "../../../example-page-factory";
+
+examplePageFactory(
+  "Snippets",
+  "Document Component",
+  () => html`<dsot-document-component-demo></dsot-document-component-demo>`
+);